fix 1 hw 17.2

+++ src/features/device/user-cases/deleteAllOtherSession.ts
@@ -1,6 +1,7 @@
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { DeviceService } from '../api/device.service';
 import { AuthService } from '../../auth/infrastructure/auth.service';
+import { UnauthorizedException } from '@nestjs/common';
 
 export class DeleteAllOtherSession {
   constructor(public refreshToken: string) {}
@@ -19,6 +20,7 @@ export class DeleteAllOtherSessionHandler
     const dataToken = await this.authService.verifyRefreshToken(
       command.refreshToken,
     );
+    if (!dataToken) throw new UnauthorizedException();
 
     return await this.deviceService.deleteAllOtherSession(
       dataToken.deviceId,
